Added functionality to start 'kkr_imp_wc' from 'GF_remote' (#101)
It is now possible to start the 'kkr_imp' workflow from a 'GF_remote'
input node instead of always having to call the 'gf_writeout' workflow.
This might be useful when you want to conduct calculations for many
different impurities while always using the same form of the impurity
cluster.

To enable this functionality the following changes have been made:

    - New (non-required) imput node 'impurity_info' has been added to the
      'kkr_imp_sub_wc'

    - Attribute 'impurity_info' has been added to '_use_methods' in the
      'kkrimp' calculation

    - If 'impurity_info' is given as input of the 'kkrimp' calculation:
      consistency check of the 'impurity_info' nodes from the input and
      the parent GF calculation

Additionaly, some minor changes to the output of 'tools_kkrimp' have been
made.