Reuse RUNOPT and RCLUSTZ in voronoi step of kkr_imp workflow
During scf a KKRhost calculation might have overwritten the RCLSUTZ and RUNOPT inputs.
When a KKRimp calculation is not done we need to make sure that the impurity potential takes over the same setting as the voronoi run from where the KKRhost calculation started.
This in particular means that the FINDSIZE run option is required to ensure that the radial mesh is the same.